recipes-bsp: u-boot: update portenta-m8 support patch for 2020.04
This patch applies the changes to the portenta-m8 support patch needed
to move to the NXP BSP u-boot 2020.04 version. The next patch will move
it to the u-boot-fio recipe.

This order was chosen so that the diff of this patch is apparent in the
git history prior to moving it.

Signed-off-by: Michael Scott <mike@foundries.io>