recipes-bsp: u-boot-fio: use script from u-boot.itb
boot.itb fit-image is now packaged as a part of the u-boot.itb so
that we can easily update it as a part of the boot software OTA
process.

Since we inherit these settings from imx8mmevk, the only change
we need here is to source the pre-loaded script at 0x44800000
instead of loading it from the boot part.

Signed-off-by: Michael Scott <mike@foundries.io>